refactor(api): Decompose code into modules
